golang+docker implementation beginnings
Uses docker for convenience in setting up all the various execution
environments. We still rely on apparmor for cap restrictions.

Includes the start of a new golang-based HTTP API and library.

This is still a work in progress, the only languages that have been updated for
the new impl are ruby, nodejs and csharp.