simpler language templates
User testing showed that solving an actual problem in the template (convert each
line to int and double) was confusing, since the code didn't apply to the actual
problem at hand.

This introduces a separate template check for most languages, with a
straightforward loop echoing stdin to stdout line-by-line, along with a comment
pointing to where you should implement your solution.

Change-Id: Iac71be27a8a2b65181292fef8e1facb769c5f180
Reviewed-on: https://gerrit.instructure.com/61953
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins