shared volume for compilation, not image commit
The image commit feels cleaner, but it's just too slow -- the `docker
commit` api call takes more than 1s on average on these containers.

So we've switched to a strategy of not committing the container, instead
loading/compiling the code onto a shared volume that gets mounted
(read-only) into each execution container.

Along with the async container removal also in this commit, this drops
the /execute response time on a simple ruby script from 1600ms to 400ms.

Change-Id: Icefd38d70bbdc25a6e1d8d0a0e49db58b4549813